Add document extraction nodes, markdown nodes, pdf-to-image, and clippy fixes
- Add extract_word, extract_pdf, extract_html nodes (33 built-in)
  - Word: zip + quick-xml parsing with metadata (Dublin Core)
  - PDF: lopdf + pdf-extract with metadata (Info dictionary)
  - HTML: html2md + ammonia with metadata (title, meta tags)
  - Format param validated: "text" or "markdown" only
  - metadata_key is opt-in to avoid duplicate context entries
- Add markdown_to_html and html_to_markdown nodes (comrak, html2md, ammonia)
- Add pdf_to_image node behind "pdf-render" feature flag (pdfium-render)
  - Supports page specs: "all", "1", "1-5", "1,3,7-9"
  - PNG/JPEG output as base64, configurable DPI
  - Honors PDFIUM_LIB_PATH env var for library discovery
- Fix clippy warnings: collapsible ifs, while_let_loop, redundant closures
- Update all docs to reflect 33+1 node count

+++ src/lua/runtime.rs
@@ -265,7 +265,7 @@ impl LuaRuntime {
 /// Convert a Lua table to serde_json::Value.
 fn lua_table_to_json(table: &LuaTable) -> Result<serde_json::Value> {
     // Check if this is an array (sequential integer keys starting from 1)
-    let len = table.len()? as i64;
+    let len = table.len()?;
     let mut is_array = len > 0;
 
     if is_array {
